Moved all project management functionality to the right panel
