+++ app/src/containers/RightContainer.tsx
@@ -1,85 +1,17 @@
 import React, { useState, useContext, useEffect, useMemo } from 'react';
+
+import { makeStyles } from '@material-ui/core/styles';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import TextField from '@material-ui/core/TextField';
-import { makeStyles } from '@material-ui/core/styles';
+import Button from '@material-ui/core/Button';
+
 import { stateContext } from '../context/context';
 import HTMLTypes from '../context/HTMLTypes';
-import Button from '@material-ui/core/Button';
 import { useHistory, withRouter } from 'react-router-dom';
-import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 
-import LoginButton from '../components/login/LoginButton';
-import SaveProjectButton from '../components/login/SaveProjectButton';
-import ProjectsFolder from '../components/login/ProjectsFolder';
-
-const useStyles = makeStyles({
-  select: {
-    fontSize: '1.25em',
-    '> .MuiSelect-icon': {
-      color: 'white'
-    }
-  },
-  selectInput: {
-    color: '#fff',
-    paddingTop: '15px',
-    paddingLeft: '15px'
-  },
-  formControl: {
-    minWidth: '125px',
-    backgroundColor: 'rgba(255,255,255,0.15)'
-  },
-  configRow: {
-    display: 'flex',
-    paddingLeft: '25px',
-    paddingRight: '25px',
-    marginTop: '20px'
-  },
-  configType: {
-    color: '#fff',
-    minWidth: '185px',
-    fontSize: '1em'
-  },
-  configValue: {
-    marginLeft: '20px'
-  },
-  buttonRow: {
-    textAlign: 'center',
-    marginTop: '25px',
-    '& > .MuiButton-textSecondary': {
-      color: 'rgba(255,0,0,0.75)'
-    }
-  },
-  button: {
-    fontSize: '1rem',
-    paddingLeft: '20px',
-    paddingRight: '20px'
-  },
-  compName: {
-    color: '#01d46d',
-    fontSize: '1.75rem'
-  },
-  configHeader: {
-    height: '70px',
-    '& > h4': {
-      fontSize: '1.25rem',
-      letterSpacing: '0.5px',
-      marginBottom: '0',
-      marginTop: '10px'
-    }
-  },
-  backButton: {
-    position: 'absolute',
-    bottom: '50px',
-    right: '125px'
-  },
-  logoutButton: {
-    position: 'absolute',
-    bottom: '50px',
-    right: '150px'
-  }
-});
+import ProjectManager from '../components/right/ProjectManager';
 
 // need to pass in props to use the useHistory feature of react router
 const RightContainer = (props): JSX.Element => {
@@ -203,7 +135,7 @@ const RightContainer = (props): JSX.Element => {
   // dispatch to 'UPDATE CSS' called when save button is clicked,
   // passing in style object constructed from all changed input values
   const handleSave = (): Object => {
-    const styleObj = {};
+    const styleObj: any = {};
     if (displayMode !== '') styleObj.display = displayMode;
     if (flexDir !== '') styleObj.flexDirection = flexDir;
     if (flexJustify !== '') styleObj.justifyContent = flexJustify;
@@ -226,222 +158,272 @@ const RightContainer = (props): JSX.Element => {
     dispatch({ type: 'DELETE CHILD', payload: {} });
   };
 
-  //
-  const handleLogout = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    e.preventDefault();
-    console.log('Logout clicked, destroying cookie, redirect to login');
-    // destroys "cookie" by clearing localStorage if guest
-    window.localStorage.clear();
-    // destroy cookie on production when not seen by chromium browser using ipcrenderer
-    window.api.delCookie();
-    // destroys cookie if user by backdating cookie expiration date
-    document.cookie = 'ssid=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;';
-    // uses useHistory to return to the login page
-    props.history.push('/login');
-  };
-
   return (
-    <div className="column right">
-      <div className={classes.configHeader}>
-        {configTarget.child ? (
-          <h4>
-            Instance of
-            {configTarget.child.type === 'component'
-              ? ' component'
-              : ' element'}{' '}
-            <br />
-            <span className={classes.compName}>{configTarget.child.name}</span>
-          </h4>
-        ) : (
-          <h4>
-            Parent component
-            <br />
-            <span className={classes.compName}>{configTarget.name}</span>
-          </h4>
-        )}
-      </div>
-      <div className={classes.configRow}>
-        <div className={classes.configType}>
-          <h3>Display:</h3>
-        </div>
-        <div className={classes.configValue}>
-          <FormControl variant="filled" className={classes.formControl}>
-            <Select
-              value={displayMode}
-              name="display"
-              onChange={handleChange}
-              displayEmpty
-              className={classes.select}
-              inputProps={{ className: classes.selectInput }}
-            >
-              <MenuItem value=""></MenuItem>
-              <MenuItem value={'block'}>block</MenuItem>
-              <MenuItem value={'inline-block'}>inline-block</MenuItem>
-              <MenuItem value={'flex'}>flex</MenuItem>
-            </Select>
-          </FormControl>
-        </div>
-      </div>
-      {/* flex options are hidden until display flex is chosen */}
-      {displayMode === 'flex' && (
+    <div className="column right ">
+      <div className="rightPanelWrapper">
         <div>
+          <div className={classes.configHeader}>
+            {configTarget.child ? (
+              <h4>
+                Instance of
+                {configTarget.child.type === 'component'
+                  ? ' component'
+                  : ' element'}{' '}
+                <br />
+                <span className={classes.compName}>
+                  {configTarget.child.name}
+                </span>
+              </h4>
+            ) : (
+              <h4>
+                Parent component
+                <br />
+                <span className={classes.compName}>{configTarget.name}</span>
+              </h4>
+            )}
+          </div>
           <div className={classes.configRow}>
             <div className={classes.configType}>
-              <h3>Flex direction:</h3>
+              <h3>Display:</h3>
             </div>
             <div className={classes.configValue}>
               <FormControl variant="filled" className={classes.formControl}>
                 <Select
-                  value={flexDir}
-                  name="flexdir"
+                  value={displayMode}
+                  name="display"
                   onChange={handleChange}
                   displayEmpty
                   className={classes.select}
                   inputProps={{ className: classes.selectInput }}
                 >
                   <MenuItem value=""></MenuItem>
-                  <MenuItem value={'row'}>row</MenuItem>
-                  <MenuItem value={'column'}>column</MenuItem>
+                  <MenuItem value={'block'}>block</MenuItem>
+                  <MenuItem value={'inline-block'}>inline-block</MenuItem>
+                  <MenuItem value={'flex'}>flex</MenuItem>
                 </Select>
               </FormControl>
             </div>
           </div>
+          {/* flex options are hidden until display flex is chosen */}
+          {displayMode === 'flex' && (
+            <div>
+              <div className={classes.configRow}>
+                <div className={classes.configType}>
+                  <h3>Flex direction:</h3>
+                </div>
+                <div className={classes.configValue}>
+                  <FormControl variant="filled" className={classes.formControl}>
+                    <Select
+                      value={flexDir}
+                      name="flexdir"
+                      onChange={handleChange}
+                      displayEmpty
+                      className={classes.select}
+                      inputProps={{ className: classes.selectInput }}
+                    >
+                      <MenuItem value=""></MenuItem>
+                      <MenuItem value={'row'}>row</MenuItem>
+                      <MenuItem value={'column'}>column</MenuItem>
+                    </Select>
+                  </FormControl>
+                </div>
+              </div>
+              <div className={classes.configRow}>
+                <div className={classes.configType}>
+                  <h3>Justify content:</h3>
+                </div>
+                <div className={classes.configValue}>
+                  <FormControl variant="filled" className={classes.formControl}>
+                    <Select
+                      value={flexJustify}
+                      name="flexjust"
+                      onChange={handleChange}
+                      displayEmpty
+                      className={classes.select}
+                      inputProps={{ className: classes.selectInput }}
+                    >
+                      <MenuItem value=""></MenuItem>
+                      <MenuItem value={'flex-start'}>flex-start</MenuItem>
+                      <MenuItem value={'flex-end'}>flex-end</MenuItem>
+                      <MenuItem value={'center'}>center</MenuItem>
+                      <MenuItem value={'space-between'}>space-between</MenuItem>
+                      <MenuItem value={'space-around'}>space-around</MenuItem>
+                      <MenuItem value={'space-evenly'}>space-evenly</MenuItem>
+                    </Select>
+                  </FormControl>
+                </div>
+              </div>
+              <div className={classes.configRow}>
+                <div className={classes.configType}>
+                  <h3>Align items:</h3>
+                </div>
+                <div className={classes.configValue}>
+                  <FormControl variant="filled" className={classes.formControl}>
+                    <Select
+                      value={flexAlign}
+                      onChange={handleChange}
+                      name="flexalign"
+                      displayEmpty
+                      className={classes.select}
+                      inputProps={{ className: classes.selectInput }}
+                    >
+                      <MenuItem value=""></MenuItem>
+                      <MenuItem value={'stretch'}>stretch</MenuItem>
+                      <MenuItem value={'flex-start'}>flex-start</MenuItem>
+                      <MenuItem value={'flex-end'}>flex-end</MenuItem>
+                      <MenuItem value={'center'}>center</MenuItem>
+                    </Select>
+                  </FormControl>
+                </div>
+              </div>
+            </div>
+          )}
           <div className={classes.configRow}>
             <div className={classes.configType}>
-              <h3>Justify content:</h3>
+              <h3>Width:</h3>
             </div>
             <div className={classes.configValue}>
               <FormControl variant="filled" className={classes.formControl}>
                 <Select
-                  value={flexJustify}
-                  name="flexjust"
+                  value={compWidth}
+                  name="width"
                   onChange={handleChange}
                   displayEmpty
                   className={classes.select}
                   inputProps={{ className: classes.selectInput }}
                 >
                   <MenuItem value=""></MenuItem>
-                  <MenuItem value={'flex-start'}>flex-start</MenuItem>
-                  <MenuItem value={'flex-end'}>flex-end</MenuItem>
-                  <MenuItem value={'center'}>center</MenuItem>
-                  <MenuItem value={'space-between'}>space-between</MenuItem>
-                  <MenuItem value={'space-around'}>space-around</MenuItem>
-                  <MenuItem value={'space-evenly'}>space-evenly</MenuItem>
+                  <MenuItem value={'auto'}>auto</MenuItem>
+                  <MenuItem value={'50%'}>50%</MenuItem>
+                  <MenuItem value={'25%'}>25%</MenuItem>
                 </Select>
               </FormControl>
             </div>
           </div>
           <div className={classes.configRow}>
             <div className={classes.configType}>
-              <h3>Align items:</h3>
+              <h3>Height:</h3>
             </div>
             <div className={classes.configValue}>
               <FormControl variant="filled" className={classes.formControl}>
                 <Select
-                  value={flexAlign}
+                  value={compHeight}
+                  name="height"
                   onChange={handleChange}
-                  name="flexalign"
                   displayEmpty
                   className={classes.select}
                   inputProps={{ className: classes.selectInput }}
                 >
                   <MenuItem value=""></MenuItem>
-                  <MenuItem value={'stretch'}>stretch</MenuItem>
-                  <MenuItem value={'flex-start'}>flex-start</MenuItem>
-                  <MenuItem value={'flex-end'}>flex-end</MenuItem>
-                  <MenuItem value={'center'}>center</MenuItem>
+                  <MenuItem value={'auto'}>auto</MenuItem>
+                  <MenuItem value={'100%'}>100%</MenuItem>
+                  <MenuItem value={'50%'}>50%</MenuItem>
                 </Select>
               </FormControl>
             </div>
           </div>
-        </div>
-      )}
-      <div className={classes.configRow}>
-        <div className={classes.configType}>
-          <h3>Width:</h3>
-        </div>
-        <div className={classes.configValue}>
-          <FormControl variant="filled" className={classes.formControl}>
-            <Select
-              value={compWidth}
-              name="width"
-              onChange={handleChange}
-              displayEmpty
-              className={classes.select}
-              inputProps={{ className: classes.selectInput }}
+          <div className={classes.configRow}>
+            <div className={classes.configType}>
+              <h3>Background color:</h3>
+            </div>
+            <div className={classes.configValue}>
+              <FormControl variant="filled" className={classes.formControl}>
+                <TextField
+                  variant="filled"
+                  name="bgcolor"
+                  className={classes.select}
+                  inputProps={{ className: classes.selectInput }}
+                  value={BGColor}
+                  onChange={handleChange}
+                />
+              </FormControl>
+            </div>
+          </div>
+          <div className={classes.buttonRow}>
+            <Button
+              color="primary"
+              className={classes.button}
+              onClick={handleSave}
             >
-              <MenuItem value=""></MenuItem>
-              <MenuItem value={'auto'}>auto</MenuItem>
-              <MenuItem value={'50%'}>50%</MenuItem>
-              <MenuItem value={'25%'}>25%</MenuItem>
-            </Select>
-          </FormControl>
+              SAVE
+            </Button>
+          </div>
+          {configTarget.child ? (
+            <div className={classes.buttonRow}>
+              <Button
+                color="secondary"
+                className={classes.button}
+                onClick={handleDelete}
+              >
+                DELETE INSTANCE
+              </Button>
+            </div>
+          ) : (
+            ''
+          )}
         </div>
-      </div>
-      <div className={classes.configRow}>
-        <div className={classes.configType}>
-          <h3>Height:</h3>
-        </div>
-        <div className={classes.configValue}>
-          <FormControl variant="filled" className={classes.formControl}>
-            <Select
-              value={compHeight}
-              name="height"
-              onChange={handleChange}
-              displayEmpty
-              className={classes.select}
-              inputProps={{ className: classes.selectInput }}
-            >
-              <MenuItem value=""></MenuItem>
-              <MenuItem value={'auto'}>auto</MenuItem>
-              <MenuItem value={'100%'}>100%</MenuItem>
-              <MenuItem value={'50%'}>50%</MenuItem>
-            </Select>
-          </FormControl>
-        </div>
-      </div>
-      <div className={classes.configRow}>
-        <div className={classes.configType}>
-          <h3>Background color:</h3>
-        </div>
-        <div className={classes.configValue}>
-          <FormControl variant="filled" className={classes.formControl}>
-            <TextField
-              variant="filled"
-              name="bgcolor"
-              className={classes.select}
-              inputProps={{ className: classes.selectInput }}
-              value={BGColor}
-              onChange={handleChange}
-            />
-          </FormControl>
-        </div>
-      </div>
-      <div className={classes.buttonRow}>
-        <Button color="primary" className={classes.button} onClick={handleSave}>
-          SAVE
-        </Button>
-      </div>
-      {configTarget.child ? (
-        <div className={classes.buttonRow}>
-          <Button
-            color="secondary"
-            className={classes.button}
-            onClick={handleDelete}
-          >
-            DELETE INSTANCE
-          </Button>
-        </div>
-      ) : (
-        ''
-      )}
-      <div className={classes.logoutButton}>
-        <SaveProjectButton />
-        <ProjectsFolder />
-        <LoginButton />
+        <ProjectManager />
       </div>
     </div>
   );
 };
+
+const useStyles = makeStyles({
+  select: {
+    fontSize: '1.25em',
+    '> .MuiSelect-icon': {
+      color: 'white'
+    }
+  },
+  selectInput: {
+    color: '#fff',
+    paddingTop: '15px',
+    paddingLeft: '15px'
+  },
+  formControl: {
+    minWidth: '125px',
+    backgroundColor: 'rgba(255,255,255,0.15)'
+  },
+  configRow: {
+    display: 'flex',
+    paddingLeft: '25px',
+    paddingRight: '25px',
+    marginTop: '20px'
+  },
+  configType: {
+    color: '#fff',
+    minWidth: '185px',
+    fontSize: '1em'
+  },
+  configValue: {
+    marginLeft: '20px'
+  },
+  buttonRow: {
+    textAlign: 'center',
+    marginTop: '25px',
+    '& > .MuiButton-textSecondary': {
+      color: 'rgba(255,0,0,0.75)'
+    }
+  },
+  button: {
+    fontSize: '1rem',
+    paddingLeft: '20px',
+    paddingRight: '20px'
+  },
+  compName: {
+    color: '#01d46d',
+    fontSize: '1.75rem'
+  },
+  configHeader: {
+    height: '70px',
+    '& > h4': {
+      fontSize: '1.25rem',
+      letterSpacing: '0.5px',
+      marginBottom: '0',
+      marginTop: '10px'
+    }
+  },
+
+});
 
 export default withRouter(RightContainer);
