+++ app/src/helperFunctions/projectGetSave.ts
@@ -5,7 +5,7 @@ export const getProjects = (): Promise<Object> => {
   let userId = window.localStorage.getItem('ssid');
   //console.log('userId from localStorage is', userId);
   const body = JSON.stringify({ userId });
-  const projects = fetch('https://localhost:8081/getProjects', {
+  const projects = fetch('http://localhost:8081/getProjects', {
     method: 'POST',
     headers: {
       'content-type': 'application/json'
@@ -33,7 +33,7 @@ export const saveProject = (
     project: workspace,
     userId: window.localStorage.getItem('ssid')
   });
-  const project = fetch('https://localhost:8081/saveProject', {
+  const project = fetch('http://localhost:8081/saveProject', {
     method: 'POST',
     headers: {
       'content-type': 'application/json'
