+++ app/src/public/styles/styleNew.css
@@ -54,3 +54,11 @@ h3 {
   padding-top: 35px;
   width: 420px;
 }
+
+.rightPanelWrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  width: 100%;
+}
